Example keybinding hook for clojure repl history
Adds keyboard shortcut for moving through the history of the Clojure
repl buffer.  The default keybinding is `S-<up>` and `S-<down>`.  So
this shortcut means you don't have to press shift to use the up and down
arrow keys.

Example of how to add a keybinding for a mode using the add-hook
function, which loads the keybinding configuration into the mode when
that mode is loaded into Emacs.

As most modes, especially Clojure, are only loaded into Emacs when
needed (eg. you open a Clojure file) the add-hook is required for this
keybinding to be used once the mode is loaded.