Refactor Org config to evaluate after Org-mode is loaded
The org-mode in Spacemacs is that downloaded from Elpa and so is not loaded by
default.  Org-mode is only loaded once the first org-mode file is opened in
Spacemacs.

To ensure the org-mode configurations are only set once org-mode is loaded, the
`with-eval-after-load` function is used to wrap each configuration.