Update: switch to org-re-reveal
org-reveal causes several issues with org-mode version 9.2 and org-reveal has
also been abandoned by its maintainer.

However, another maintainer forked org-reveal and created org-re-reveal that
works fine with org-mode 9.2.

It does require an additional package, as org-re-reveal is not (yet) included in
the org layer (I think there is a Pull Request for this, but probably caught up
in the freeze for Spacemacs 0.300).