Removed clojure-lint package - dropping Joker linter
Moving to the clj-kondo linter tool instead of Joker.

I will be using clj-kondo as it uses Clojure and so should provide better
coverage of the language.  There is much more community activity in clj-kondo,
especially around the linting rules it is apply.

clj-kondo has sufficiently progressed that joker linter is no longer required.