Named default layout to Global, display to true
The default layout is a global layout, in that it can access all the buffers
opened using `SPC b b`.  Therefore the layout has been named Global to make this
very clear.

Set the name of the default layout to display in modeline, again making it
explicit as to which layout you are in.