Change to which-key keybinding ordering
Spacemacs has changed the ordering of keybindings within which-key.  This
commit includes a configuration to restore the original keybinding order.  The
configuration is commented.

The new keybindings are organized as follows:

The lower-case keybindings are shown first, in alphabetical order.
These are typically the most used functions.

Next are the upper-case keybindings in alphabetical order, typically lesser used
alternatives to the lower-case keybindings.

Finally sub-menus are shown in alphabetical order.