Over-ride SPC / keybinding to provide helm-swoop
The default command provided by `SPC /` does not provide a simple way to end the
search.  It highlights all matches and they stay highlighted until `SPC /` is
pressed again and the previous search pattern is deleted or changed to something
that has no matches.

`SPC /` also causes issues with evil editing if `RTN` is not pressed after the
search pattern.

The configuration over-rides the default search tool and replaces it with the
recommended `helm-swoop` tool.

Helm swoop also supports the search options and uses ripgrep or ag (when
installed).