Clojure lsp server with clj-kondo - commented
Add the configuration required to run an LSP server for Clojure using the
clj-kondo-lsp-server-standalone.jar

Added as a comment as not actively used.

Only needed for environments where the clj-kondo binary does not work (windows,
graal) as it has the same features as including clj-kondo as a linter in the
Clojure layer variables.