anakondo static analysis for Clojure - commented
anakondo statically analyses Clojure projects providing auto-completion without
running a Clojure REPL.

anakondo is a very new package and is not part of Spacemacs.  It can be included
via dotspacemacs-additional-packages and manually started with `SPC SPC
anakondo-minor-mode`

Remove the comments from the `autoload` and `add-hook` lines to automatically
analyse a Clojure project when it is open.  Be aware that this will analyse all
clojure projects that are loaded as layers after restarting Emacs.