Org-journal heading with checkbox snippet
A common action is to create a new level 2 heading with checkboxes.

No key binding has been found to create a checkbox list item, so this snippet
removed the need to type - [ ] to create the first checkbox.

Snippet also adds a cookie to the title to track completed check boxes.