Magit Forge - configure number of open and closed topics
A topic is an issue or pull request and the list of each can be configured to
display a number of open and closed items.

The first number is the open items, the second is the closed items.  0 means no
items will be displayed.

The configuration will show 100 open topics and not show any closed topics, for
both issues and pull requests.